Unpack the gradle distributions without restoring timestamps from the archive
Fixes #210

The timestamp in the archive may be far in the past (reproducible archives - https://docs.gradle.org/9.2.0/userguide/working_with_files.html#sec:reproducible_archives)
and may cause problems in situations where 32bit applications interact with the filesystem.
As the timestampp is entirely irrelevant, resetting it to the uncompress time avoids the problem.

Signed-off-by: Christoph Obexer <cobexer@gradle.com>